:nail_care: style: UserListWrapper component naming
